Update mod list on profile activate and remove
I had forgotten to write the code such that the Factorio mod list was updated whenever a profile was swapped or deleting, causing the profile to get out of sync with the mod list and causing unwanted behavior. Fixed.\nAlso made it so that when the installed mods were polled, it would include the base mod to keep consistency.