Beginning of mod portal integration
Set up basic api call to acquire all mods currently hosted on Factorio
mod portal. Simply displays the names on the table at this time.
Checks the first box on #15

Small bug I have yet to figure out. When switching to the onlineMods
page, it shows the list of profiles in the table. Probably due to me
using the same id, but there doesn't appear to be a way for me to clear
it before the mod information is loaded. At least, not that I've found
yet.