Allow updating of mods
Can now update a mod. It first downloads the updated mod to the
directory (had error when attempting to save it to a data folder first,
haven't figured it out), then deletes the old one.
Uses client-side logic to check whether a mod can be updated or not, no
checks built into server-side.