Adapt onlineMods.js for new mod structure
Edited existing functionality so the full mod list will be shown
correctly.

Had to stop the mod list being updated during fetching while
page_onlineMods was being viewed, as it slowed the application down too
much. Will have to come up with a more optimized way of handling it if I
want that behavior.

Can select the mod versions by a simple dropdown menu in the table.

Currently breaks the existing functionality of checking mod
updates/downloads, but in all honesty it wasn't that great to begin
with. Will have to rewrite a few portions to get that working properly.
Next commit.