Allow downloading specific releases
Reworked the download method in the Mod Manager to get the specific
release selected.
Edited the installed mods page to find the latest compatible update and
provide a quick link to download that.

To keep from rewriting and having too much duplicate functionality right
now, disabled downloading an installed mods from the online mods page,
as that works fine in the appropriate page.

This appears to have broken the profile auto-updating feature, will fix
that in next commit.