More comprehensive library and scripts
Include webpack for compilation of jsx code. Added config for webpack
and adjusted scripts to account for it, included additional test/bundle
step in all packaging.

Added in a helper file for tests for testing the React code, and
adjusted timeout limit for standard testing, since it often takes longer
on the first run and tests will now need to pass before I'll be allowed
to package.