Fill out Profiles component and logic
Now have it replicating the functionality of the profiles section of the
app previously.

Made a change to the file structure of the app. Separated components
into their own folders to keep things contained better, and let the
master components holde styling and subcomponents that they need without
cluttering things up.