Added in Footer component and logic
Implemented the footer section.

Left out the online mod download indicator due to indecision about how
to handle it in React (may leave it out of footer and have overlay
notifications). It'll be easier once online mods component and data is
hooked up.

My structure is beginning to seem a bit silly (a logic file named
footer.js?) but I'll keep it for now and worry about restructing later
once I have everything in place. Got to learn when to stop refactoring,
someday.