Connected together Main and Renderer process
Added middlware and methods to manage communication between the
processes (processi?)

Tweaked deleteMod method of ModManager since it was easy to do index,
but left ProfileManager along, although I would benefit from switching
to using activeProfile as an index value.

Removed requestDownloadMiddleware as its function was served the same as
the new middleware to pass messages to Main process.