Add some logging in dev-runner process for testing
That's all the extra testing stuff I will add. Finally found the issue 
was with one of the development libraries, they would cause the app 
never to run again after a first successful run. Not sure what's up with 
that, but I don't care enough about devtools to bother with it.
Leaving out debug libraries such as electron-debug, devtools, and 
vue-devtools.